+++ CHANGELOG.md
@@ -1377,3 +1377,4 @@
 - [2026-05-24 09:25] feat(cache): integrate offline caching for cache state
 - [2026-05-24 09:28] fix(storage): fix memory leak by cancelling storage subscription
 - [2026-05-24 09:53] feat(storage): add bi-directional event bus handler for storage
+- [2026-05-24 09:55] chore(navigation): update dependencies and cleanup unused imports in navigation